Read the filter value in BookingFilterListViewModel
